Add TPMS sensor mapping to vehicle config
Moved vehicle configuration handler to a new file and added a TPMS tab for sensor mapping in the web interface. Updated OvmsVehicle to load TPMS wheel index mapping from config, and added corresponding member variable. Also fixed climate schedule help text example formatting.
